+++ R/aaa.r
@@ -1,0 +1,2 @@
+utils::globalVariables(c("station_values", "date_start", "date_end", "keyword", "network", "date_range",
+                         "station", "show", "show_date", "word", "snippet"))