+++ NEWS.md
@@ -1,2 +1,8 @@
+0.3.0
+* `top_text()` returns a tidy data frame by default
+
+0.2.0
+* Some extra helper functions
+
 0.1.0 
 * Initial release